refactor(ajudae-person-select): extract image to another component
