+++ app/src/components/AjudaePersonSelect/index.js
@@ -1,39 +1,21 @@
 import React, { useEffect, useState } from 'react';
+import AjudaePersonImage from '../AjudaePersonImage';
 
-import personIcon1 from '../../assets/person_1.png';
-import personIcon2 from '../../assets/person_2.png';
-import personIcon3 from '../../assets/person_3.png';
-import personIcon4 from '../../assets/person_4.png';
-import { Container, PersonItem, PersonImage } from './styles';
+import { Container } from './styles';
 
 const AjudaePersonSelect = ({ defaultPersonImage, onSelectPersonImage }) => {
-  const items = [
-    {
-      color: '#FFC542',
-      source: personIcon1,
-      value: 'personIcon1',
-    },
-    {
-      color: '#3DD598',
-      source: personIcon2,
-      value: 'personIcon2',
-    },
-    {
-      color: '#FF575F',
-      source: personIcon3,
-      value: 'personIcon3',
-    },
-    {
-      color: '#755FE2',
-      source: personIcon4,
-      value: 'personIcon4',
-    },
-  ];
   const [selectedPersonIndex, setSelectedPersonIndex] = useState(-1);
+
+  const selectableImages = [
+    'personIcon1',
+    'personIcon2',
+    'personIcon3',
+    'personIcon4',
+  ];
 
   useEffect(() => {
     setSelectedPersonIndex(
-      items.findIndex((item) => item.value === defaultPersonImage),
+      selectableImages.findIndex((image) => image === defaultPersonImage),
     );
 
     console.log(defaultPersonImage);
@@ -47,14 +29,13 @@ const AjudaePersonSelect = ({ defaultPersonImage, onSelectPersonImage }) => {
 
   return (
     <Container>
-      {items.map(({ color, source, value }, index) => (
-        <PersonItem
-          key={index}
-          color={color}
+      {selectableImages.map((personImage, index) => (
+        <AjudaePersonImage
+          key={personImage}
+          personImage={personImage}
           selected={index === selectedPersonIndex}
-          onPress={() => onPressPerson(index, color, value)}>
-          <PersonImage source={source} />
-        </PersonItem>
+          onPress={(image) => onPressPerson(index, image.color, image.value)}
+        />
       ))}
     </Container>
   );
